fix(gateway): cast InsertAsync result to int after ORM long widening
Dawning.ORM.Dapper b285c1f widened InsertAsync from int to long. All gateway entities use GUID PKs, so the inserted-id return value is always 0 (no auto-key); the long->int cast is a no-op behaviorally and avoids cascading the type change through ValueTask<int> repository/service interfaces and ApiResponse<int> controllers.

Updates 20 repositories under Dawning.Identity.Infra.Data/Repository to apply (int) cast at the boundary.

Build green; gateway tests 199/200 (the one failure is a pre-existing performance flake unrelated to ORM).

+++ apps/gateway/src/Dawning.Identity.Infra.Data/Repository/Administration/ClaimTypeRepository.cs
@@ -103,7 +103,7 @@ namespace Dawning.Identity.Infra.Data.Repository.Administration
             }
 
             ClaimTypeEntity entity = model.ToEntity();
-            return await _context.Connection.InsertAsync(entity, _context.Transaction);
+            return (int)await _context.Connection.InsertAsync(entity, _context.Transaction);
         }
 
         /// <summary>
