+++ apps/gateway/src/Dawning.Identity.Infra.Data/Repository/Administration/UserRoleRepository.cs
@@ -99,7 +99,7 @@ namespace Dawning.Identity.Infra.Data.Repository.Administration
                 CreatedBy = operatorId,
             };
 
-            var result = await _context.Connection.InsertAsync(entity, _context.Transaction);
+            var result = (int)await _context.Connection.InsertAsync(entity, _context.Transaction);
             return result > 0;
         }
