+++ apps/gateway/src/Dawning.Identity.Infra.Data/Repository/Monitoring/AlertRuleRepository.cs
@@ -69,7 +69,7 @@ public class AlertRuleRepository : IAlertRuleRepository
         entity.CreatedAt = DateTime.UtcNow;
         entity.UpdatedAt = DateTime.UtcNow;
 
-        var result = await _context.Connection.InsertAsync(entity, _context.Transaction);
+        var result = (int)await _context.Connection.InsertAsync(entity, _context.Transaction);
         return result;
     }
