+++ apps/gateway/src/Dawning.Identity.Infra.Data/Repository/OpenIddict/TokenRepository.cs
@@ -147,7 +147,7 @@ namespace Dawning.Identity.Infra.Data.Repository.OpenIddict
             }
 
             TokenEntity entity = model.ToEntity();
-            return await _context.Connection.InsertAsync(entity, _context.Transaction);
+            return (int)await _context.Connection.InsertAsync(entity, _context.Transaction);
         }
 
         /// <summary>
